Rename the `zio-kafka-test-utils` module to `zio-kafka-testkit` to make it more explicit that it's something that can be used by zio-kafka users
Also, change the base namespace of this module from `zio.kafka` to `zio.kafka.testkit` to make things more explicit
