+++ zio-kafka-bench/src/main/scala/zio/kafka/bench/ConsumersComparisonBenchmark.scala
@@ -4,13 +4,13 @@ import io.github.embeddedkafka.EmbeddedKafka
 import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.apache.kafka.common.serialization.ByteArrayDeserializer
 import org.openjdk.jmh.annotations._
-import zio.kafka.KafkaTestUtils.{ consumerSettings, produceMany, producer, simpleConsumer }
 import zio.kafka.admin.AdminClient.TopicPartition
 import zio.kafka.bench.ZioBenchmark.randomThing
 import zio.kafka.consumer.{ Consumer, ConsumerSettings, Subscription }
-import zio.kafka.embedded.Kafka
 import zio.kafka.producer.Producer
 import zio.kafka.serde.Serde
+import zio.kafka.testkit.KafkaTestUtils.{ consumerSettings, produceMany, producer, simpleConsumer }
+import zio.kafka.testkit.embedded.Kafka
 import zio.{ durationInt, ULayer, ZIO, ZLayer }
 
 import java.util.concurrent.TimeUnit
@@ -38,7 +38,7 @@ class ConsumersComparisonBenchmark extends ZioBenchmark[Env] {
   private val kafkaConsumer: ZLayer[ConsumerSettings, Throwable, LowLevelKafka] =
     ZLayer.scoped {
       ZIO.acquireRelease {
-        ZIO.service[ConsumerSettings].flatMap { settings =>
+        ZIO.serviceWithZIO[ConsumerSettings] { settings =>
           ZIO.attemptBlocking {
             new KafkaConsumer[Array[Byte], Array[Byte]](
               settings.driverSettings.asJava,
