+++ zio-kafka-testkit/src/main/resources/truststore/ca-key
