+++ zio-kafka-testkit/src/main/resources/truststore/kafka.truststore.jks
