+++ zio-kafka-testkit/src/main/scala/zio/kafka/testkit/KafkaRandom.scala
@@ -1,4 +1,4 @@
-package zio.kafka
+package zio.kafka.testkit
 
 import zio.{ Task, ZIO }
