+++ zio-kafka-testkit/src/main/scala/zio/kafka/testkit/KafkaTestUtils.scala
@@ -1,4 +1,4 @@
-package zio.kafka
+package zio.kafka.testkit
 
 import org.apache.kafka.clients.consumer.{ ConsumerConfig, ConsumerRecord }
 import org.apache.kafka.clients.producer.{ ProducerRecord, RecordMetadata }
@@ -7,9 +7,9 @@ import zio.kafka.admin._
 import zio.kafka.consumer.Consumer.{ AutoOffsetStrategy, OffsetRetrieval }
 import zio.kafka.consumer._
 import zio.kafka.consumer.diagnostics.Diagnostics
-import zio.kafka.embedded.Kafka
 import zio.kafka.producer._
 import zio.kafka.serde.{ Deserializer, Serde, Serializer }
+import zio.kafka.testkit.embedded.Kafka
 
 import java.io.File
 import java.nio.file.{ Files, StandardCopyOption }
