+++ zio-kafka-testkit/src/main/scala/zio/kafka/testkit/ZIOSpecWithKafka.scala
@@ -1,11 +1,11 @@
-package zio.kafka
+package zio.kafka.testkit
 
-import zio.kafka.embedded.Kafka
 import zio.test._
 import zio.ZLayer
+import zio.kafka.testkit.embedded.Kafka
 
 /**
- * This should be used if you want to run your tests with a shared-across-your-suites embedded Kafka cluster.
+ * This trait should be used if you want to run your tests with a shared-across-your-suites embedded Kafka cluster.
  *
  * More information about sharing a resource between multiple suites can be found here:
  * https://zio.dev/reference/test/sharing-layers-between-multiple-files/
